+++ .eslintrc.js
@@ -185,6 +185,11 @@ module.exports = {
             position: "after",
           },
           {
+            pattern: "~*",
+            group: "internal",
+            position: "before",
+          },
+          {
             pattern: "~*/**",
             group: "internal",
             position: "before",
@@ -274,6 +279,7 @@ module.exports = {
       { props: "never", children: "never", propElementValues: "always" },
     ],
     "react/prop-types": "off",
+    "react/display-name": "off", // displayName isn't needed in most cases as it's inferred from variable name and it's breaking types and storybook doc gen
     "space-before-function-paren": [
       "error",
       {
@@ -298,6 +304,7 @@ module.exports = {
       files: [
         "storybook/*",
         "**/_docs/**/*",
+        "**/__tests__/**/*",
         "*.stories.tsx",
         "*.spec.ts",
         "*.spec.tsx",