+++ packages/components/jest.config.ts
@@ -6,6 +6,8 @@ const jestConfig: JestConfigWithTsJest = {
   roots: ["<rootDir>"],
   moduleNameMapper: {
     ...sharedConfig.moduleNameMapper,
+    "~tests": "<rootDir>/__tests__/index",
+    "~tests/(.*)$": "<rootDir>/__tests__/$1",
     "~types/(.*)$": "<rootDir>/src/types/$1",
     "~utils/(.*)$": "<rootDir>/src/utils/$1",
     "~components/(.*)$": "<rootDir>/src/$1",