+++ packages/components/src/Filter/FilterDatePicker/utils/validateDate.spec.ts
@@ -1,3 +1,5 @@
+import { waitFor, screen } from "@testing-library/react"
+import { renderWithIntl } from "~tests"
 import { validateDate } from "./validateDate"
 
 describe("validateDate()", () => {
@@ -19,70 +21,85 @@ describe("validateDate()", () => {
     expect(newDate).toBeUndefined()
   })
 
-  it("returns expected response when selected day is invalid and input value is undefined", () => {
+  it("returns expected response when selected day is invalid and input value is undefined", async () => {
     const date = new Date("potato")
     const inputValue = undefined
     const { validationResponse, newDate } = validateDate({ date, inputValue })
 
-    expect(validationResponse).toStrictEqual({
+    expect(newDate).toBeUndefined()
+
+    const { validationMessage, ...response } = validationResponse
+
+    expect(response).toStrictEqual({
       date,
       inputValue,
-      validationMessage: {
-        status: "error",
-        message: "Date is invalid",
-      },
       isInvalid: true,
       isDisabled: false,
       isEmpty: false,
       isValidDate: false,
     })
 
-    expect(newDate).toBeUndefined()
+    expect(validationMessage?.status).toBe("error")
+
+    renderWithIntl(validationMessage?.message)
+    await waitFor(() => {
+      expect(screen.getByText("Date is invalid")).toBeVisible()
+    })
   })
 
-  it("returns expected response when selected day is invalid and input value is empty string", () => {
+  it("returns expected response when selected day is invalid and input value is empty string", async () => {
     const date = new Date("potato")
     const inputValue = ""
     const { validationResponse, newDate } = validateDate({ date, inputValue })
 
-    expect(validationResponse).toStrictEqual({
+    expect(newDate).toBeUndefined()
+
+    const { validationMessage, ...response } = validationResponse
+
+    expect(response).toStrictEqual({
       date,
       inputValue,
-      validationMessage: {
-        status: "error",
-        message: "Date is invalid",
-      },
       isInvalid: true,
       isDisabled: false,
       isEmpty: false,
       isValidDate: false,
     })
 
-    expect(newDate).toBeUndefined()
+    expect(validationMessage?.status).toBe("error")
+
+    renderWithIntl(validationMessage?.message)
+    await waitFor(() => {
+      expect(screen.getByText("Date is invalid")).toBeVisible()
+    })
   })
 
-  it("returns expected response when selected day is invalid and input value is valid string", () => {
+  it("returns expected response when selected day is invalid and input value is valid string", async () => {
     const date = new Date("potato")
     const inputValue = "potato"
     const { validationResponse, newDate } = validateDate({ date, inputValue })
 
-    expect(validationResponse).toStrictEqual({
+    expect(newDate).toBeUndefined()
+
+    const { validationMessage, ...response } = validationResponse
+
+    expect(response).toStrictEqual({
       date,
       inputValue,
-      validationMessage: {
-        status: "error",
-        message: "potato is an invalid date",
-      },
       isInvalid: true,
       isDisabled: false,
       isEmpty: false,
       isValidDate: false,
     })
 
-    expect(newDate).toBeUndefined()
+    expect(validationMessage?.status).toBe("error")
+
+    renderWithIntl(validationMessage?.message)
+    await waitFor(() => {
+      expect(screen.getByText("potato is an invalid date")).toBeVisible()
+    })
   })
 
-  it("returns expected response when selected day is disabled", () => {
+  it("returns expected response when selected day is disabled", async () => {
     const date = new Date("2022-03-01")
     const inputValue = "03/01/2022"
     const disabledDays = [new Date("2022-03-01")]
@@ -92,20 +109,27 @@ describe("validateDate()", () => {
       disabledDays,
     })
 
-    expect(validationResponse).toStrictEqual({
+    expect(newDate).toBeUndefined()
+
+    const { validationMessage, ...response } = validationResponse
+
+    expect(response).toStrictEqual({
       date,
       inputValue,
-      validationMessage: {
-        status: "error",
-        message: "03/01/2022 is not available, try another date",
-      },
       isInvalid: false,
       isDisabled: true,
       isEmpty: false,
       isValidDate: false,
     })
 
-    expect(newDate).toBeUndefined()
+    expect(validationMessage?.status).toBe("error")
+
+    renderWithIntl(validationMessage?.message)
+    await waitFor(() => {
+      expect(
+        screen.getByText("03/01/2022 is not available, try another date")
+      ).toBeVisible()
+    })
   })
 
   it("returns expected response when selected day is valid", () => {