+++ packages/components/src/__overlays__/v2.ts
@@ -1,0 +1,2 @@
+// Alias of current
+export * from "./Tooltip/v2"